File Content on frontend matches the backend. Editing files is now possible
